tests: Refactoring fullstack tests packages Related to #100
